refactor: hook방식으로 바꾼다
